requirements.txt
langchain
langchain-anthropic
langchain-community
sentence-transformers
faiss-cpu
pypdf
streamlit
python-dotenv